Midi Device code reorganization
