+++ internal/pkg/midi/event.go
@@ -2,10 +2,14 @@ package midi
 
 import (
 	"fmt"
+	"os"
+	"strings"
 
 	"github.com/gethiox/HIDI/internal/pkg/logger"
 	"github.com/gethiox/HIDI/internal/pkg/midi/config"
 )
+
+var log = logger.GetLogger()
 
 const (
 	// message types
@@ -29,7 +33,7 @@ const (
 	TimingStop     uint8 = 0b11111100
 )
 
-var intervalToString = map[int]string{
+var IntervalToString = map[int]string{
 	0:  "Perfect unison",
 	1:  "Minor second",
 	2:  "Major second",
@@ -191,4 +195,37 @@ func ExtractEvents(d []byte) ([]Event, []byte) {
 	}
 	leftover = append(leftover, data[start:]...)
 	return events, leftover
+}
+
+func DetectDevices() []IODevice {
+	fd, err := os.Open("/dev/snd")
+	if err != nil {
+		panic(err)
+	}
+	entries, err := fd.ReadDir(0)
+	if err != nil {
+		panic(err)
+	}
+
+	var devices = make([]IODevice, 0)
+
+	for _, entry := range entries {
+		if entry.IsDir() {
+			continue
+		}
+
+		if strings.HasPrefix(entry.Name(), "midi") {
+			devices = append(devices, IODevice{path: fmt.Sprintf("/dev/snd/%s", entry.Name())})
+		}
+	}
+
+	return devices
+}
+
+type IODevice struct {
+	path string
+}
+
+func (d *IODevice) Open() (*os.File, error) {
+	return os.OpenFile(d.path, os.O_RDWR|os.O_SYNC, 0)
 }
