+++ internal/pkg/midi/leds.go
@@ -1,124 +1,1 @@
 package midi
-
-import "github.com/holoplot/go-evdev"
-
-func init() {
-	for k, v := range KeyToLedName {
-		LedNameToKey[v] = k
-	}
-}
-
-var LedNameToKey = map[string]evdev.EvCode{} // filled up with init()
-
-var KeyToLedName = map[evdev.EvCode]string{ // hardware button to OpenRGB LED name mapping
-	evdev.KEY_ESC:       "Key: Escape",
-	evdev.KEY_GRAVE:     "Key: `",
-	evdev.KEY_TAB:       "Key: Tab",
-	evdev.KEY_CAPSLOCK:  "Key: Caps Lock",
-	evdev.KEY_LEFTSHIFT: "Key: Left Shift",
-	evdev.KEY_LEFTCTRL:  "Key: Left Control",
-	// evdev.KEY_:           "Key: \\ (ISO)", // ?
-	evdev.KEY_1:          "Key: 1",
-	evdev.KEY_Q:          "Key: Q",
-	evdev.KEY_A:          "Key: A",
-	evdev.KEY_Z:          "Key: Z",
-	evdev.KEY_LEFTMETA:   "Key: Left Windows",
-	evdev.KEY_F1:         "Key: F1",
-	evdev.KEY_2:          "Key: 2",
-	evdev.KEY_W:          "Key: W",
-	evdev.KEY_S:          "Key: S",
-	evdev.KEY_X:          "Key: X",
-	evdev.KEY_LEFTALT:    "Key: Left Alt",
-	evdev.KEY_F2:         "Key: F2",
-	evdev.KEY_3:          "Key: 3",
-	evdev.KEY_E:          "Key: E",
-	evdev.KEY_D:          "Key: D",
-	evdev.KEY_C:          "Key: C",
-	evdev.KEY_F3:         "Key: F3",
-	evdev.KEY_4:          "Key: 4",
-	evdev.KEY_R:          "Key: R",
-	evdev.KEY_F:          "Key: F",
-	evdev.KEY_V:          "Key: V",
-	evdev.KEY_F4:         "Key: F4",
-	evdev.KEY_5:          "Key: 5",
-	evdev.KEY_T:          "Key: T",
-	evdev.KEY_G:          "Key: G",
-	evdev.KEY_B:          "Key: B",
-	evdev.KEY_SPACE:      "Key: Space",
-	evdev.KEY_F5:         "Key: F5",
-	evdev.KEY_6:          "Key: 6",
-	evdev.KEY_Y:          "Key: Y",
-	evdev.KEY_H:          "Key: H",
-	evdev.KEY_N:          "Key: N",
-	evdev.KEY_F6:         "Key: F6",
-	evdev.KEY_7:          "Key: 7",
-	evdev.KEY_U:          "Key: U",
-	evdev.KEY_J:          "Key: J",
-	evdev.KEY_M:          "Key: M",
-	evdev.KEY_F7:         "Key: F7",
-	evdev.KEY_8:          "Key: 8",
-	evdev.KEY_I:          "Key: I",
-	evdev.KEY_K:          "Key: K",
-	evdev.KEY_COMMA:      "Key: ,",
-	evdev.KEY_RIGHTALT:   "Key: Right Alt",
-	evdev.KEY_F8:         "Key: F8",
-	evdev.KEY_9:          "Key: 9",
-	evdev.KEY_O:          "Key: O",
-	evdev.KEY_L:          "Key: L",
-	evdev.KEY_DOT:        "Key: .",
-	evdev.KEY_COMPOSE:    "Key: Menu",
-	evdev.KEY_F9:         "Key: F9",
-	evdev.KEY_0:          "Key: 0",
-	evdev.KEY_P:          "Key: P",
-	evdev.KEY_SEMICOLON:  "Key: ;",
-	evdev.KEY_SLASH:      "Key: /",
-	evdev.KEY_RIGHTMETA:  "Key: Right Windows",
-	evdev.KEY_F10:        "Key: F10",
-	evdev.KEY_MINUS:      "Key: -",
-	evdev.KEY_LEFTBRACE:  "Key: [",
-	evdev.KEY_APOSTROPHE: "Key: '",
-	evdev.KEY_F11:        "Key: F11",
-	evdev.KEY_EQUAL:      "Key: =",
-	evdev.KEY_RIGHTBRACE: "Key: ]",
-	// evdev.KEY_:             "Key: #", // ?
-	evdev.KEY_F12:          "Key: F12",
-	evdev.KEY_BACKSPACE:    "Key: Backspace",
-	evdev.KEY_BACKSLASH:    "Key: \\ (ANSI)",
-	evdev.KEY_ENTER:        "Key: Enter",
-	evdev.KEY_RIGHTSHIFT:   "Key: Right Shift",
-	evdev.KEY_RIGHTCTRL:    "Key: Right Control",
-	evdev.KEY_SYSRQ:        "Key: Print Screen",
-	evdev.KEY_INSERT:       "Key: Insert",
-	evdev.KEY_DELETE:       "Key: Delete",
-	evdev.KEY_LEFT:         "Key: Left Arrow",
-	evdev.KEY_SCROLLLOCK:   "Key: Scroll Lock",
-	evdev.KEY_HOME:         "Key: Home",
-	evdev.KEY_END:          "Key: End",
-	evdev.KEY_UP:           "Key: Up Arrow",
-	evdev.KEY_DOWN:         "Key: Down Arrow",
-	evdev.KEY_PAUSE:        "Key: Pause/Break",
-	evdev.KEY_PAGEUP:       "Key: Page Up",
-	evdev.KEY_PAGEDOWN:     "Key: Page Down",
-	evdev.KEY_RIGHT:        "Key: Right Arrow",
-	evdev.KEY_NUMLOCK:      "Key: Num Lock",
-	evdev.KEY_KP7:          "Key: Number Pad 7",
-	evdev.KEY_KP4:          "Key: Number Pad 4",
-	evdev.KEY_KP1:          "Key: Number Pad 1",
-	evdev.KEY_KP0:          "Key: Number Pad 0",
-	evdev.KEY_KPSLASH:      "Key: Number Pad /",
-	evdev.KEY_KP8:          "Key: Number Pad 8",
-	evdev.KEY_KP5:          "Key: Number Pad 5",
-	evdev.KEY_KP2:          "Key: Number Pad 2",
-	evdev.KEY_MUTE:         "Key: Media Mute",
-	evdev.KEY_KPASTERISK:   "Key: Number Pad *",
-	evdev.KEY_KP9:          "Key: Number Pad 9",
-	evdev.KEY_KP6:          "Key: Number Pad 6",
-	evdev.KEY_KP3:          "Key: Number Pad 3",
-	evdev.KEY_KPDOT:        "Key: Number Pad .",
-	evdev.KEY_PREVIOUSSONG: "Key: Media Previous",
-	evdev.KEY_KPMINUS:      "Key: Number Pad -",
-	evdev.KEY_KPPLUS:       "Key: Number Pad +",
-	evdev.KEY_PLAYPAUSE:    "Key: Media Play/Pause",
-	evdev.KEY_NEXTSONG:     "Key: Media Next",
-	evdev.KEY_KPENTER:      "Key: Number Pad Enter",
-}
